Can deploy wiki pages as children of README.
Generate multiple HTMLs from Markdowns found under src/wiki
Synchronize output directory with children of your remote README. This means
that if you remote a wiki page, it will be deleted on confluence on next deploy.
Title of pages is the discrimator used to check the existance of a page.